requirements.txt
pandas
psycopg2-binary
python-dotenv
sqlalchemy
scikit-learn
matplotlib
jupyter
